Split translation tests to allow a bit quicker execution
